added quant conv v2 and activation
